# Conflicts:
#	.idea/workspace.xml
#	src/main/java/application/Index.java
#	src/main/java/controller/MainMenuController.java
#	src/main/java/controller/SubjectCategoryController.java
#	src/main/java/controller/Test.java
#	src/main/java/controller/UpdatePersonalInformationController.java
#	src/main/java/dao/AccountDAO.java
#	src/main/java/dao/DepartmentDAO.java
#	src/main/java/dao/DetailLearningChartDAO.java
#	src/main/java/dao/DetailRegisterClassDao.java
#	src/main/java/dao/DetailRegisterSubject.java
#	src/main/java/dao/DetailRegisterSubjectDao.java
#	src/main/java/dao/InfomationStudentDAO.java
#	src/main/java/dao/LearningChartDAO.java
#	src/main/java/dao/ProgramDAO.java
#	src/main/java/dao/RegisterSubjectDao.java
#	src/main/java/dao/ScheduleDAO.java
#	src/main/java/dao/StudentDao.java
#	src/main/java/dao/SubjectDao.java
#	src/main/java/dao/ToeicDao.java
#	src/main/java/dao/impl/AccountDAOImpl.java
#	src/main/java/dao/impl/DepartmentDAOImpl.java
#	src/main/java/dao/impl/DetailLearningChartDAOImpl.java
#	src/main/java/dao/impl/DetailRegisterClassDaoImpl.java
#	src/main/java/dao/impl/DetailRegisterSubjectDaoImpl.java
#	src/main/java/dao/impl/InfomationStudentDAOImpl.java
#	src/main/java/dao/impl/LearningChartDAOImpl.java
#	src/main/java/dao/impl/ProgramDAOImpl.java
#	src/main/java/dao/impl/ScheduleDAOImpl.java
#	src/main/java/dao/impl/StudentDaoImpl.java
#	src/main/java/dao/impl/SubjectDaoImpl.java
#	src/main/java/dao/impl/ToeicDaoImpl.java
#	src/main/java/dao/impl/UserDao.java
#	src/main/java/logics/AccountLogic.java
#	src/main/java/logics/DepartmentLogic.java
#	src/main/java/logics/DetailLearningChartLogic.java
#	src/main/java/logics/DetailRegisterClassLogic.java
#	src/main/java/logics/DetailSubjectLogic.java
#	src/main/java/logics/InfomationStudentLogic.java
#	src/main/java/logics/LearningChartLogic.java
#	src/main/java/logics/ProgramLogic.java
#	src/main/java/logics/RegisterSubjectLogic.java
#	src/main/java/logics/ScheduleLogic.java
#	src/main/java/logics/StudentLogic.java
#	src/main/java/logics/SubjectLogic.java
#	src/main/java/logics/ToeicLogic.java
#	src/main/java/logics/impl/AccountLogicImpl.java
#	src/main/java/logics/impl/DepartmentLogicImpl.java
#	src/main/java/logics/impl/DetailLearningChartLogicImpl.java
#	src/main/java/logics/impl/DetailRegisterClassLogicImpl.java
#	src/main/java/logics/impl/InfomationStudentLogicImpl.java
#	src/main/java/logics/impl/LearningChartLogicImpl.java
#	src/main/java/logics/impl/ProgramLogicImpl.java
#	src/main/java/logics/impl/ScheduleLogicImpl.java
#	src/main/java/logics/impl/StudentLogicImpl.java
#	src/main/java/logics/impl/SubjectLogicImpl.java
#	src/main/java/logics/impl/ToeicLogicImpl.java
#	src/main/java/utils/Constants.java
#	src/main/java/validate/ChangePasswordValidate.java
#	src/main/java/validate/DepartmentValidate.java
#	src/main/java/validate/DetailLearningChartValidate.java
#	src/main/java/validate/DetailSubjectValidate.java
#	src/main/java/validate/InfomationStudentValidate.java
#	src/main/java/validate/LearningChartValidate.java
#	src/main/java/validate/LoginValidate.java
#	src/main/java/validate/ProgramValidate.java
#	src/main/java/validate/ScheduleValidate.java
#	src/main/java/validate/StudentValidate.java
#	src/main/java/validate/ToeicValidate.java
#	src/main/resources/css/style.css
#	src/main/resources/fxml/change-password.fxml
#	src/main/resources/fxml/login.fxml
#	src/main/resources/fxml/subject-category.fxml
#	src/test/java/TestAccount.java
#	src/test/java/TestDepartment.java
#	src/test/java/TestInfomationStudent.java
#	src/test/java/dao/impl/AccountDAOImplTest.java
#	src/test/java/dao/impl/DepartmentDAOImplTest.java
#	src/test/java/dao/impl/DetailLearningChartDAOImplTest.java
#	src/test/java/dao/impl/LearningChartDAOImplTest.java
#	src/test/java/dao/impl/ScheduleDAOImplTest.java
#	src/test/java/dao/impl/StudentDaoImplTest.java

+++ src/main/java/entities/TblDetailregisterclass.java
@@ -2,7 +2,7 @@
  * Copyright(C) 2018 Bùi Trung Hiếu
  * StudentInformationSystem, entities - IntelliJ IDEA
  * TblDetailregisterclass.java
- * Created at 08:57 ~ 22/11/2018 by Bùi Trung Hiếu
+ * Created at 16:09 ~ 06/12/2018 by Bùi Trung Hiếu
  */
 package entities;
 
@@ -12,102 +12,102 @@ import java.util.Objects;
 /**
  * Description about this Class/Interface/Enum...
  *
- * @author Bùi Trung Hiếu  Created: 08:57 ~ 22/11/2018
+ * @author Bùi Trung Hiếu  Created: 16:09 ~ 06/12/2018
  */
 @Entity @Table(name = "detailregisterclass", schema = "sis_db", catalog = "")
 @IdClass(TblDetailregisterclassPK.class) public class TblDetailregisterclass {
-    private int idRegister;
-    private int idSchedule;
-    private String classCode;
-    private Integer tuition;
-    private Integer midSemPoint;
-    private Integer finalSemPoint;
-    private Integer stt;
-    private String timeModified;
-
-    @Id @Column(name = "idRegister") public int getIdRegister() {
-        return idRegister;
-    }
-
-    public void setIdRegister(int idRegister) {
-        this.idRegister = idRegister;
-    }
-
-    @Id @Column(name = "idSchedule") public int getIdSchedule() {
-        return idSchedule;
-    }
-
-    public void setIdSchedule(int idSchedule) {
-        this.idSchedule = idSchedule;
-    }
-
-    @Basic @Column(name = "classCode") public String getClassCode() {
-        return classCode;
-    }
-
-    public void setClassCode(String classCode) {
-        this.classCode = classCode;
-    }
-
-    @Basic @Column(name = "tuition") public Integer getTuition() {
-        return tuition;
-    }
-
-    public void setTuition(Integer tuition) {
-        this.tuition = tuition;
-    }
-
-    @Basic @Column(name = "midSemPoint") public Integer getMidSemPoint() {
-        return midSemPoint;
-    }
-
-    public void setMidSemPoint(Integer midSemPoint) {
-        this.midSemPoint = midSemPoint;
-    }
-
-    @Basic @Column(name = "finalSemPoint") public Integer getFinalSemPoint() {
-        return finalSemPoint;
-    }
-
-    public void setFinalSemPoint(Integer finalSemPoint) {
-        this.finalSemPoint = finalSemPoint;
-    }
-
-    @Basic @Column(name = "stt") public Integer getStt() {
-        return stt;
-    }
-
-    public void setStt(Integer stt) {
-        this.stt = stt;
-    }
-
-    @Basic @Column(name = "timeModified") public String getTimeModified() {
-        return timeModified;
-    }
-
-    public void setTimeModified(String timeModified) {
-        this.timeModified = timeModified;
-    }
-
-    @Override public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        TblDetailregisterclass that = (TblDetailregisterclass) o;
-        return idRegister == that.idRegister &&
-               idSchedule == that.idSchedule &&
-               Objects.equals(classCode, that.classCode) &&
-               Objects.equals(tuition, that.tuition) &&
-               Objects.equals(midSemPoint, that.midSemPoint) &&
-               Objects.equals(finalSemPoint, that.finalSemPoint) &&
-               Objects.equals(stt, that.stt) &&
-               Objects.equals(timeModified, that.timeModified);
-    }
-
-    @Override public int hashCode() {
-        return Objects.hash(idRegister, idSchedule, classCode, tuition, midSemPoint, finalSemPoint, stt, timeModified);
-    }
+	private int idRegister;
+	private int idSchedule;
+	private String classCode;
+	private Integer tuition;
+	private Double midSemPoint;
+	private Double finalSemPoint;
+	private Integer stt;
+	private String timeModified;
+	
+	@Id @Column(name = "idRegister") public int getIdRegister() {
+		return idRegister;
+	}
+	
+	public void setIdRegister(int idRegister) {
+		this.idRegister = idRegister;
+	}
+	
+	@Id @Column(name = "idSchedule") public int getIdSchedule() {
+		return idSchedule;
+	}
+	
+	public void setIdSchedule(int idSchedule) {
+		this.idSchedule = idSchedule;
+	}
+	
+	@Basic @Column(name = "classCode") public String getClassCode() {
+		return classCode;
+	}
+	
+	public void setClassCode(String classCode) {
+		this.classCode = classCode;
+	}
+	
+	@Basic @Column(name = "tuition") public Integer getTuition() {
+		return tuition;
+	}
+	
+	public void setTuition(Integer tuition) {
+		this.tuition = tuition;
+	}
+	
+	@Basic @Column(name = "midSemPoint") public Double getMidSemPoint() {
+		return midSemPoint;
+	}
+	
+	public void setMidSemPoint(Double midSemPoint) {
+		this.midSemPoint = midSemPoint;
+	}
+	
+	@Basic @Column(name = "finalSemPoint") public Double getFinalSemPoint() {
+		return finalSemPoint;
+	}
+	
+	public void setFinalSemPoint(Double finalSemPoint) {
+		this.finalSemPoint = finalSemPoint;
+	}
+	
+	@Basic @Column(name = "stt") public Integer getStt() {
+		return stt;
+	}
+	
+	public void setStt(Integer stt) {
+		this.stt = stt;
+	}
+	
+	@Basic @Column(name = "timeModified") public String getTimeModified() {
+		return timeModified;
+	}
+	
+	public void setTimeModified(String timeModified) {
+		this.timeModified = timeModified;
+	}
+	
+	@Override public boolean equals(Object o) {
+		if (this == o) {
+			return true;
+		}
+		if (o == null || getClass() != o.getClass()) {
+			return false;
+		}
+		TblDetailregisterclass that = (TblDetailregisterclass) o;
+		return idRegister == that.idRegister &&
+		       idSchedule == that.idSchedule &&
+		       Objects.equals(classCode, that.classCode) &&
+		       Objects.equals(tuition, that.tuition) &&
+		       Objects.equals(midSemPoint, that.midSemPoint) &&
+		       Objects.equals(finalSemPoint, that.finalSemPoint) &&
+		       Objects.equals(stt, that.stt) &&
+		       Objects.equals(timeModified, that.timeModified);
+	}
+	
+	@Override public int hashCode() {
+		return Objects.hash(idRegister, idSchedule, classCode, tuition, midSemPoint, finalSemPoint, stt, timeModified);
+	}
 }
